refactor to remove dependency on PartialState for parallelism config
